Remove ember-glimmer-template-compiler package.
It is no longer needed. All implementation has been moved back into
ember-template-compiler itself.
