+++ lib/packages.js
@@ -14,8 +14,18 @@ module.exports = function(features) {
     'ember-template-compiler': {
       trees: null,
       templateCompilerOnly: true,
-      requirements: ['container', 'ember-metal', 'ember-environment', 'ember-console', 'ember-glimmer-template-compiler'],
-      templateCompilerVendor: ['simple-html-tokenizer', 'backburner']
+      requirements: ['container', 'ember-metal', 'ember-environment', 'ember-console'],
+      templateCompilerVendor: [
+        'simple-html-tokenizer',
+        'backburner',
+        'glimmer-wire-format',
+        'glimmer-syntax',
+        'glimmer-util',
+        'glimmer-compiler',
+        'glimmer-reference',
+        'glimmer-runtime',
+        'handlebars'
+      ]
     },
     'ember-routing':              { trees: null,  vendorRequirements: ['router', 'route-recognizer'],
                                     requirements: ['ember-runtime', 'ember-views'] },
@@ -36,21 +46,6 @@ module.exports = function(features) {
         'glimmer-node'
       ],
       testingVendorRequirements: []
-    },
-
-    'ember-glimmer-template-compiler': {
-      trees: null,
-      templateCompilerOnly: true,
-      requirements: [],
-      templateCompilerVendor: [
-        'glimmer-wire-format',
-        'glimmer-syntax',
-        'glimmer-util',
-        'glimmer-compiler',
-        'glimmer-reference',
-        'glimmer-runtime',
-        'handlebars'
-      ]
     }
   };
