+++ packages/ember-application/tests/system/application_test.js
@@ -13,7 +13,7 @@ import EmberObject from 'ember-runtime/system/object';
 import { setSearchDisabled as setNamespaceSearchDisabled } from 'ember-runtime/system/namespace';
 import EmberRoute from 'ember-routing/system/route';
 import jQuery from 'ember-views/system/jquery';
-import { compile } from 'ember-template-compiler/tests/utils/helpers';
+import { compile } from 'ember-template-compiler';
 import { _loaded } from 'ember-runtime/system/lazy_load';
 import { getDebugFunction, setDebugFunction } from 'ember-metal/debug';
 import { setTemplates, setTemplate } from 'ember-glimmer';
