+++ packages/ember-application/tests/system/dependency_injection/custom_resolver_test.js
@@ -2,7 +2,7 @@ import jQuery from 'ember-views/system/jquery';
 import run from 'ember-metal/run_loop';
 import Application from 'ember-application/system/application';
 import DefaultResolver from 'ember-application/system/resolver';
-import { compile } from 'ember-template-compiler/tests/utils/helpers';
+import { compile } from 'ember-template-compiler';
 
 let application;
