--- packages/ember-glimmer-template-compiler/lib/index.js
@@ -1,4 +1,0 @@
-export { default as compile } from './system/compile';
-export { default as precompile } from './system/precompile';
-export { default as defaultCompileOptions } from './system/compile-options';
-export { registerPlugin } from './system/compile-options';
