--- packages/ember-template-compiler/lib/compiler.js
@@ -1,5 +1,0 @@
-import * as compiler from 'ember-glimmer-template-compiler';
-
-export default function pickCompiler() {
-  return compiler;
-}
