+++ packages/ember-template-compiler/lib/plugins/index.js
@@ -7,6 +7,11 @@ import TransformInlineLinkTo from './transform-inline-link-to';
 import TransformOldClassBindingSyntax from './transform-old-class-binding-syntax';
 import DeprecateRenderModel from './deprecate-render-model';
 import AssertReservedNamedArguments from './assert-reserved-named-arguments';
+import TransformActionSyntax from './transform-action-syntax';
+import TransformInputTypeSyntax from './transform-input-type-syntax';
+import TransformAttrsIntoArgs from './transform-attrs-into-args';
+import TransformEachInIntoEach from './transform-each-in-into-each';
+import TransformHasBlockSyntax from './transform-has-block-syntax';
 
 export default Object.freeze([
   TransformOldBindingSyntax,
@@ -17,5 +22,10 @@ export default Object.freeze([
   TransformInlineLinkTo,
   TransformOldClassBindingSyntax,
   DeprecateRenderModel,
-  AssertReservedNamedArguments
+  AssertReservedNamedArguments,
+  TransformActionSyntax,
+  TransformInputTypeSyntax,
+  TransformAttrsIntoArgs,
+  TransformEachInIntoEach,
+  TransformHasBlockSyntax
 ]);
