+++ packages/ember-template-compiler/lib/plugins/transform-action-syntax.js
