+++ packages/ember-template-compiler/lib/plugins/transform-attrs-into-args.js
