+++ packages/ember-template-compiler/lib/plugins/transform-has-block-syntax.js
