+++ packages/ember-template-compiler/lib/plugins/transform-input-type-syntax.js
