--- packages/ember-template-compiler/lib/system/register-plugin.js
@@ -1,5 +1,0 @@
-import compiler from '../compiler';
-
-let { registerPlugin } = compiler();
-
-export default registerPlugin;
