+++ packages/ember-template-compiler/tests/plugins/deprecate-render-model-test.js
@@ -1,4 +1,4 @@
-import { compile } from '../utils/helpers';
+import { compile } from 'ember-template-compiler';
 
 QUnit.module('ember-template-compiler: deprecate-model-render');
 
@@ -15,4 +15,3 @@ QUnit.test('Using `{{render` with model provides a deprecation', function() {
     });
   }, expectedMessage);
 });
-
