+++ packages/ember-template-compiler/tests/plugins/transform-inline-link-to-test.js
@@ -1,4 +1,4 @@
-import { compile } from '../utils/helpers';
+import { compile } from 'ember-template-compiler';
 
 QUnit.module('ember-template-compiler: assert-no-view-and-controller-paths without legacy view support');
