+++ packages/ember-template-compiler/tests/system/compile_options_test.js
@@ -1,16 +1,16 @@
-import { defaultCompileOptions } from 'ember-template-compiler';
-import defaultPlugins from 'ember-template-compiler/plugins';
+import { compileOptions } from 'ember-template-compiler';
+import { defaultPlugins } from 'ember-template-compiler';
 
 QUnit.module('ember-template-compiler: default compile options');
 
 QUnit.test('default options are a new copy', function() {
-  notEqual(defaultCompileOptions(), defaultCompileOptions());
+  notEqual(compileOptions(), compileOptions());
 });
 
 QUnit.test('has default AST plugins', function(assert) {
   assert.expect(defaultPlugins.length);
 
-  let plugins = defaultCompileOptions().plugins.ast;
+  let plugins = compileOptions().plugins.ast;
 
   for (let i = 0; i < defaultPlugins.length; i++) {
     let plugin = defaultPlugins[i];
