--- packages/ember-template-compiler/tests/utils/helpers.js
@@ -1,1 +1,0 @@
-export { compile } from 'ember-glimmer-template-compiler/tests/utils/helpers';
