+++ packages/ember-testing/tests/helpers_test.js
@@ -13,7 +13,7 @@ import setupForTesting from 'ember-testing/setup_for_testing';
 import EmberRouter from 'ember-routing/system/router';
 import EmberRoute from 'ember-routing/system/route';
 import EmberApplication from 'ember-application/system/application';
-import { compile } from 'ember-template-compiler/tests/utils/helpers';
+import { compile } from 'ember-template-compiler';
 
 import {
   pendingRequests,
