+++ packages/ember/tests/application_lifecycle_test.js
@@ -3,7 +3,7 @@ import Route from 'ember-routing/system/route';
 import run from 'ember-metal/run_loop';
 import { Component, setTemplates, getTemplates } from 'ember-glimmer';
 import jQuery from 'ember-views/system/jquery';
-import { compile } from 'ember-template-compiler/tests/utils/helpers';
+import { compile } from 'ember-template-compiler';
 import controllerFor from 'ember-routing/system/controller_for';
 
 let App, TEMPLATES, appInstance, router;
