+++ packages/ember/tests/controller_test.js
@@ -1,7 +1,7 @@
 import Controller from 'ember-runtime/controllers/controller';
 import Route from 'ember-routing/system/route';
 import run from 'ember-metal/run_loop';
-import { compile } from 'ember-template-compiler/tests/utils/helpers';
+import { compile } from 'ember-template-compiler';
 import Application from 'ember-application/system/application';
 import { Component, setTemplates, setTemplate } from 'ember-glimmer';
 import jQuery from 'ember-views/system/jquery';
