+++ packages/ember/tests/integration/multiple-app-test.js
@@ -1,5 +1,5 @@
 import run from 'ember-metal/run_loop';
-import { compile } from 'ember-template-compiler/tests/utils/helpers';
+import { compile } from 'ember-template-compiler';
 import Application from 'ember-application/system/application';
 import { Component } from 'ember-glimmer';
 import jQuery from 'ember-views/system/jquery';
