+++ packages/ember/tests/view_instrumentation_test.js
@@ -2,7 +2,7 @@ import run from 'ember-metal/run_loop';
 import $ from 'ember-views/system/jquery';
 import Application from 'ember-application/system/application';
 import { subscribe, reset } from 'ember-metal/instrumentation';
-import { compile } from 'ember-template-compiler/tests/utils/helpers';
+import { compile } from 'ember-template-compiler';
 import { setTemplates, setTemplate } from 'ember-glimmer';
 
 let App, $fixture;
